+++ src/Core/OperationCacheContext.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace JacyImp\ApiPlatformOperationCache\Core;
 
+use ApiPlatform\Metadata\HttpOperation;
 use JacyImp\ApiPlatformOperationCache\Metadata\OperationCache;
 
 /**
@@ -12,6 +13,7 @@ use JacyImp\ApiPlatformOperationCache\Metadata\OperationCache;
 final readonly class OperationCacheContext
 {
     public function __construct(
+        public HttpOperation $operation,
         public OperationCache $cache,
         public string $key,
     ) {